Rename: change history directory structure
