+++ app/mypage/[id]/(history)/@header/postHistory/page.tsx
@@ -1,0 +1,6 @@
+import Header from '@/app/_components/common/header/page'
+import Back from '@/app/_components/common/header/_components/Back'
+
+export default function PostHistory() {
+  return <Header leftSection={<Back />} title={'내가 작성한 글'} />
+}
