+++ app/mypage/[id]/(history)/history.module.scss
