--- app/mypage/[id]/histPreview/histPreview.module.scss
@@ -1,7 +1,0 @@
-.container {
-  width: 100%;
-  padding: 12px 16px;
-  display: flex;
-  flex-direction: column;
-  gap: 8px;
-}
