+++ app/mypage/[id]/nickNameModify/page.tsx
@@ -1,0 +1,11 @@
+import Header from '@/app/_components/common/header/page'
+import Back from '@/app/_components/common/header/_components/Back'
+
+export default function NickNameModify() {
+  return (
+    <>
+      <Header leftSection={<Back />} title={'닉네임 수정'} />
+      <div>nick setting</div>
+    </>
+  )
+}
