--- app/mypage/[id]/nickNameSetting/page.tsx
@@ -1,3 +1,0 @@
-export default function NickNameSetting() {
-  return <div>nick setting</div>
-}
